Add the very first opera documentation
In this commit we are paving the way for establishing the initial
documentation for xOpera orchestrator. The documentation is currently
in the form of .rst files and is prepared to be build with Read the
Docs and Sphinx documentation tool generator.

With these changes we set the structure for hosting our docs on GitHub
Pages. To accomplish that we need to separate source and build files
into two folders. We prepared a Makefile to be able to build our
documentation locally. And last but not least, there is a .nojekyll
file we use to bypass Jekyll processing on GitHub Pages.

This first documentation is not perfect and is meant to be a tool
introduction page or microsite for European Horizon 2020 RADON project
Currently the generated HTML files are located in build folder which is
definitely not the best solution. For the future we could create an
account for Read the Docs and connect it to this repo so that we won't
have to build and push these files manually anymore.