Remove svg screencasts from CLI docs
The svg images that are showing xOpera commands are not so good after
all because its hard to follow the whole orchestration workflow. We
will rather keep just the examples that can be explored slowly and
copied anytime. The svg images are also not so practical when building
documentation PDFs, so removing them will bring us some benefits there
too.