Use sequential CI/CD workflows instead of parallel
Within this commit we will fix out CI/CD configuration that is
currently failing. We have two GitHub Actions workflows - first for
building the docs and artifacts and the second for publishing docs to
GitHub pages. Since the second one uses the artifacts from the first,
the second has to wait for the first to complete.